# @main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eduman Api Documentation",
        "description": "Eduman Api Documentation",
        "contact": {
            "name": "Md Nasir Uddin",
            "email": "nasir.chalo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/setting/menu/by-category": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get list of Menus",
                "description": "Get list of Menus",
                "operationId": "getMenusByCategory",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Header_Home_1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}